+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 All notable changes to this project will be documented in this file.
 
+<!-- git-cliff-unreleased-start -->
+## 1.9.3 - **not yet released**
+
+### 🐛 Bug Fixes
+
+- **sitemaps:** Allow Sitemap.from_xml_string to opt into enqueue host filtering ([#2162](https://github.com/apify/crawlee-python/pull/2162)) ([a3c20c1](https://github.com/apify/crawlee-python/commit/a3c20c15a15f7ffa5644640adde6daf1dd4c3523)) by [@aryansk](https://github.com/aryansk), closes [#2118](https://github.com/apify/crawlee-python/issues/2118)
+
+
+<!-- git-cliff-unreleased-end -->
 ## [1.9.2](https://github.com/apify/crawlee-python/releases/tag/v1.9.2) (2026-08-17)
 
 ### 🐛 Bug Fixes